tests/wpt/web-platform-tests/XMLHttpRequest/responseurl.html

<!doctype html>
<html>
  <head>
    <title>XMLHttpRequest: responseURL test</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <link rel="help" href="https://xhr.spec.whatwg.org/#the-responseurl-attribute"/>
  </head>
  <body>
    <div id="log"></div>
    <script>
      test(function() {
        var client = new XMLHttpRequest()
        assert_equals(client.responseURL, "")
        client.open("GET", "foo.html", false)
        client.send()
        expected = location.href.replace(/[^/]*$/, 'foo.html')
        assert_equals(client.responseURL, expected)
      })
      test(function() {
        var client = new XMLHttpRequest()
        assert_equals(client.responseURL, "")
        expected = "resources/image.gif"
        client.open("GET", "resource/redirect.py&location=" + expected, false)
        client.send()
        assert_equals(client.responseURL, expected)
      })
    </script>
  </body>
</html>
